lazyload fixes & image optimizations
