+++ client/src/scenes/widgets/FeedWidget.jsx
@@ -82,6 +82,7 @@ const FeedWidget = ({ limit = 10 }) => {
               likes,
               comments,
               isEdited,
+              aspectRatio,
             }) => (
               <PostWidget
                 key={_id}
@@ -98,6 +99,7 @@ const FeedWidget = ({ limit = 10 }) => {
                 likes={likes}
                 comments={comments}
                 isEdited={isEdited}
+                aspectRatio={aspectRatio}
               />
             )
           )}
